Fix UI issues with Actions
