+++ AuroraEditor/.githash
@@ -1,1 +1,1 @@
-67cf7bedd6f30a247be5701c7978b5d4fd907a4e
+769ad8ed05a33308b514df7c312c882537f1b48e
