--- AuroraEditor/Base/Source Control/UI/Actions/WorkflowRunsView.swift
@@ -1,47 +1,0 @@
-//
-//  WorkflowRunsView.swift
-//  AuroraEditor
-//
-//  Created by Nanashi Li on 2022/09/13.
-//  Copyright © 2022 Aurora Company. All rights reserved.
-//
-
-import SwiftUI
-
-struct WorkflowRunsView: View {
-
-    @ObservedObject
-    private var actionsModel: GitHubActions
-
-    @State
-    private var workflowId: String
-
-    init(workspace: WorkspaceDocument, workflowId: String) {
-        self.actionsModel = .init(workspace: workspace)
-        self.workflowId = workflowId
-    }
-
-    var body: some View {
-        VStack {
-            NavigationView {
-                VStack {
-                    Divider()
-                        .padding()
-                    List(actionsModel.workflowRuns, id: \.id) { run in
-                        NavigationLink {
-                            WorkflowJobsView(workspace: actionsModel.workspace,
-                                             runId: String(run.id),
-                                             jobName: run.name)
-                        } label: {
-                            WorkflowRunCell(workflowRun: run)
-                        }
-                    }
-                    .frame(minWidth: 450)
-                }
-            }
-        }
-        .onAppear {
-            actionsModel.fetchWorkflowRuns(workflowId: workflowId)
-        }
-    }
-}
