+++ AuroraEditor/Base/Version Control/UI/Branch/CreateNewBranchView.swift
