+++ AuroraEditor/Base/Version Control/UI/Project History/Cell/CommitHistoryCellView.swift
